+++ modules/cli_commands/load.js
@@ -1,0 +1,6 @@
+module.exports = {
+   // command: 'load',
+   disabled: true,
+   alias: [],
+   callback: () => {}
+}